Moved some ship functions to ship control. Simplified some code.
