+++ EDWayPoint.py
@@ -24,8 +24,9 @@ Author: sumzer0@yahoo.com
 
 
 class EDWayPoint:
-    def __init__(self, ed_ap, is_odyssey=True):
+    def __init__(self, ed_ap, cb, is_odyssey=True):
         self.ap = ed_ap
+        self.ap_ckb = cb
         # self.is_odyssey = is_odyssey
         self.filename = ''
         self.stats_log = {'Colonisation': 0, 'Construction': 0, 'Fleet Carrier': 0, 'Station': 0}
@@ -65,11 +66,11 @@ class EDWayPoint:
             self.num_waypoints = len(self.waypoints)
             self.filename = filename
             self.ap.config['WaypointFilepath'] = filename
-            self.ap.ap_ckb('log', f"Loaded Waypoint file: {filename}")
+            self.ap_ckb('log', f"Loaded Waypoint file: {filename}")
             logger.debug("EDWayPoint: read json:" + str(ss))
             return True
 
-        self.ap.ap_ckb('log', f"Waypoint file is invalid. Check log file for details.")
+        self.ap_ckb('log', f"Waypoint file is invalid. Check log file for details.")
         return False
 
     def _read_waypoints(self, filename='./waypoints/waypoints.json'):
@@ -85,7 +86,7 @@ class EDWayPoint:
             # Perform any checks on the data returned
             # Check if the waypoint data contains the 'GlobalShoppingList' (new requirement)
             if 'GlobalShoppingList' not in s:
-                # self.ap.ap_ckb('log', f"Waypoint file is invalid. Check log file for details.")
+                # self.ap_ckb('log', f"Waypoint file is invalid. Check log file for details.")
                 logger.warning(f"Waypoint file {filename} is invalid or old version. "
                                f"It does not contain a 'GlobalShoppingList' waypoint.")
                 s = None
@@ -229,10 +230,10 @@ class EDWayPoint:
         # From the journal, this works for stations (incl. outpost), colonisation ship and megaships
         if ap.jn.ship_state()['StationServices'] is not None:
             if 'commodities' not in ap.jn.ship_state()['StationServices']:
-                self.ap.ap_ckb('log', f"No commodities market at docked location.")
+                self.ap_ckb('log', f"No commodities market at docked location.")
                 return
         else:
-            self.ap.ap_ckb('log', f"No station services at docked location.")
+            self.ap_ckb('log', f"No station services at docked location.")
             return
 
         # Determine type of station we are at
@@ -242,12 +243,12 @@ class EDWayPoint:
             if station_type == StationType.ColonisationShip:
                 # Colonisation Ship
                 self.stats_log['Colonisation'] = self.stats_log['Colonisation'] + 1
-                self.ap.ap_ckb('log', f"Executing trade with Colonisation Ship.")
+                self.ap_ckb('log', f"Executing trade with Colonisation Ship.")
                 logger.debug(f"Execute Trade: On Colonisation Ship")
             if station_type == StationType.SpaceConstructionDepot:
                 # Construction Ship
                 self.stats_log['Construction'] = self.stats_log['Construction'] + 1
-                self.ap.ap_ckb('log', f"Executing trade with Orbital Construction Ship.")
+                self.ap_ckb('log', f"Executing trade with Orbital Construction Ship.")
                 logger.debug(f"Execute Trade: On Orbital Construction Site")
 
             # Go to station services
@@ -273,7 +274,7 @@ class EDWayPoint:
 
         else:
             # Regular Station or Fleet Carrier or Squadron Carrier in Buy/Sell mode
-            self.ap.ap_ckb('log', "Executing trade.")
+            self.ap_ckb('log', "Executing trade.")
             logger.debug(f"Execute Trade: On Regular Station")
             self.stats_log['Station'] = self.stats_log['Station'] + 1
 
@@ -302,7 +303,7 @@ class EDWayPoint:
             #     ap.keys.send('UI_Down')
             #     ap.keys.send('UI_Select')  # Select Commodities
 
-            self.ap.ap_ckb('log+vce', "Downloading commodities data from market.")
+            self.ap_ckb('log+vce', "Downloading commodities data from market.")
 
             # Wait for market to update
             market_time_new = ""
@@ -470,11 +471,11 @@ class EDWayPoint:
         also can then perform trades if specific in the waypoints file.
         """
         if len(self.waypoints) == 0:
-            self.ap.ap_ckb('log+vce', "No Waypoint file loaded. Exiting Waypoint Assist.")
+            self.ap_ckb('log+vce', "No Waypoint file loaded. Exiting Waypoint Assist.")
             return
 
         self.step = 0  # start at first waypoint
-        self.ap.ap_ckb('log', "Waypoint file: " + str(Path(self.filename).name))
+        self.ap_ckb('log', "Waypoint file: " + str(Path(self.filename).name))
         self.reset_stats()
 
         # Loop until complete, or error
@@ -507,7 +508,7 @@ class EDWayPoint:
                     break
 
             if dest_key is None:
-                self.ap.ap_ckb('log+vce', "Waypoint list has been completed.")
+                self.ap_ckb('log+vce', "Waypoint list has been completed.")
                 break
 
             # Is this a new waypoint?
@@ -528,7 +529,7 @@ class EDWayPoint:
             next_wp_station = next_waypoint.get('StationName', '').upper()
 
             if new_waypoint:
-                self.ap.ap_ckb('log+vce', f"Next Waypoint: {next_wp_station} in {next_wp_system}")
+                self.ap_ckb('log+vce', f"Next Waypoint: {next_wp_station} in {next_wp_system}")
 
             # ====================================
             # Target and travel to a System
@@ -537,7 +538,7 @@ class EDWayPoint:
             # Check current system and go to next system if different and not blank
             if next_wp_system == "" or (cur_star_system == next_wp_system):
                 if new_waypoint:
-                    self.ap.ap_ckb('log+vce', f"Already in target System.")
+                    self.ap_ckb('log+vce', f"Already in target System.")
             else:
                 # Check if the current nav route is to the target system
                 last_nav_route_sys = self.ap.nav_route.get_last_system().upper()
@@ -546,16 +547,16 @@ class EDWayPoint:
                 if ((last_nav_route_sys == next_wp_system) and
                         (destination_body == 0 and destination_name != "")):
                     # No need to target system
-                    self.ap.ap_ckb('log+vce', f"System already targeted.")
+                    self.ap_ckb('log+vce', f"System already targeted.")
                 else:
-                    self.ap.ap_ckb('log+vce', f"Targeting system {next_wp_system}.")
+                    self.ap_ckb('log+vce', f"Targeting system {next_wp_system}.")
                     # Select destination in galaxy map based on name
                     res = self.ap.galaxy_map.set_gal_map_destination_text(self.ap, next_wp_system,
                                                                           self.ap.jn.ship_state)
                     if res:
-                        self.ap.ap_ckb('log', f"System has been targeted.")
+                        self.ap_ckb('log', f"System has been targeted.")
                     else:
-                        self.ap.ap_ckb('log+vce', f"Unable to target {next_wp_system} in Galaxy Map.")
+                        self.ap_ckb('log+vce', f"Unable to target {next_wp_system} in Galaxy Map.")
                         _abort = True
                         break
 
@@ -564,10 +565,10 @@ class EDWayPoint:
                 keys.send('TargetNextRouteSystem')
 
                 # Jump to the destination system
-                self.ap.ap_ckb('log+vce', f"Jumping to {next_wp_system}.")
+                self.ap_ckb('log+vce', f"Jumping to {next_wp_system}.")
                 res = self.ap.jump_to_system(scr_reg)
                 if not res:
-                    self.ap.ap_ckb('log', f"Failed to jump to {next_wp_system}.")
+                    self.ap_ckb('log', f"Failed to jump to {next_wp_system}.")
                     _abort = True
                     break
 
@@ -584,25 +585,25 @@ class EDWayPoint:
                 # Check if we are at the correct station. Note that for FCs, the station name
                 # reported by the Journal is only the ship identifier (ABC-123) and not the carrier name.
                 # So we need to check if the ID (ABC-123) is at the end of the target ('Fleety McFleet ABC-123').
-                self.ap.ap_ckb('log', f"Docked at {cur_station}.")
-                self.ap.ap_ckb('log', f"Checking station name against target station: {next_wp_station}.")
+                self.ap_ckb('log', f"Docked at {cur_station}.")
+                self.ap_ckb('log', f"Checking station name against target station: {next_wp_station}.")
                 if cur_station_type == StationType.FleetCarrier:
                     if next_wp_station.endswith(cur_station):
                         docked_at_stn = True
                     else:
-                        self.ap.ap_ckb('log', f"FleetCarrier name: {next_wp_station} does not end with: {cur_station}.")
+                        self.ap_ckb('log', f"FleetCarrier name: {next_wp_station} does not end with: {cur_station}.")
 
                 elif cur_station_type == StationType.SquadronCarrier:
                     if next_wp_station.endswith(cur_station):
                         docked_at_stn = True
                     else:
-                        self.ap.ap_ckb('log', f"SquadronCarrier name: {next_wp_station} does not end with: {cur_station}.")
+                        self.ap_ckb('log', f"SquadronCarrier name: {next_wp_station} does not end with: {cur_station}.")
 
                 elif 'System Colonisation Ship'.upper() in next_wp_station:
                     if cur_station_type == StationType.ColonisationShip:
                         docked_at_stn = True
                     else:
-                        self.ap.ap_ckb('log', f"Current station is not a Colonisation Ship.")
+                        self.ap_ckb('log', f"Current station is not a Colonisation Ship.")
 
                 # elif next_wp_station.startswith('Orbital Construction Site'.upper()):
                 #     if (cur_station_type == 'SurfaceStation'.upper() and
@@ -612,24 +613,24 @@ class EDWayPoint:
                     docked_at_stn = True
 
                 if not docked_at_stn:
-                    self.ap.ap_ckb('log', f"We are not docked at the target station.")
+                    self.ap_ckb('log', f"We are not docked at the target station.")
 
             # Check current station and go to it if different
             if docked_at_stn:
                 if new_waypoint:
-                    self.ap.ap_ckb('log+vce', f"Already at target Station: {next_wp_station}.")
+                    self.ap_ckb('log+vce', f"Already at target Station: {next_wp_station}.")
             else:
                 # Check if we need to travel to a station, else we are done.
                 # This may be by 1) System bookmark, 2) Galaxy bookmark or 3) by Station Name text
                 if sys_bookmark or gal_bookmark or next_wp_station != "":
                     # If waypoint file has a Station Name associated then attempt targeting it
-                    self.ap.ap_ckb('log+vce', f"Targeting Station: {next_wp_station}")
+                    self.ap_ckb('log+vce', f"Targeting Station: {next_wp_station}")
 
                     if gal_bookmark:
                         # Set destination via gal bookmark, not system bookmark
                         res = self.ap.galaxy_map.set_gal_map_dest_bookmark(self.ap, gal_bookmark_type, gal_bookmark_num)
                         if not res:
-                            self.ap.ap_ckb('log+vce', f"Unable to set Galaxy Map bookmark.")
+                            self.ap_ckb('log+vce', f"Unable to set Galaxy Map bookmark.")
                             _abort = True
                             break
 
@@ -637,7 +638,7 @@ class EDWayPoint:
                         # Set destination via system bookmark
                         res = self.ap.system_map.set_sys_map_dest_bookmark(self.ap, sys_bookmark_type, sys_bookmark_num)
                         if not res:
-                            self.ap.ap_ckb('log+vce', f"Unable to set System Map bookmark.")
+                            self.ap_ckb('log+vce', f"Unable to set System Map bookmark.")
                             _abort = True
                             break
 
@@ -645,12 +646,12 @@ class EDWayPoint:
                         # Set destination via system name
                         res = self.ap.nav_panel.lock_destination(next_wp_station)
                         if not res:
-                            self.ap.ap_ckb('log+vce', f"Unable to set Nav Panel OCR bookmark.")
+                            self.ap_ckb('log+vce', f"Unable to set Nav Panel OCR bookmark.")
                             _abort = True
                             break
 
                     else:
-                        self.ap.ap_ckb('log+vce', f"No bookmark defined.")
+                        self.ap_ckb('log+vce', f"No bookmark defined.")
                         _abort = True
                         break
 
@@ -659,7 +660,7 @@ class EDWayPoint:
                     sleep(1)  # Allow status log to update
                     continue
                 else:
-                    self.ap.ap_ckb('log+vce', f"Arrived at target System: {next_wp_system}")
+                    self.ap_ckb('log+vce', f"Arrived at target System: {next_wp_system}")
 
             # ====================================
             # Dock and Trade at Station
@@ -668,20 +669,20 @@ class EDWayPoint:
             # Are we at the correct station to trade?
             if docked_at_stn:  # and (next_wp_station != "" or sys_bookmark):
                 # Docked - let do trade
-                self.ap.ap_ckb('log+vce', f"Execute trade at Station: {next_wp_station}")
+                self.ap_ckb('log+vce', f"Execute trade at Station: {next_wp_station}")
                 self.execute_trade(self.ap, dest_key)
 
             # Mark this waypoint as completed
             self.mark_waypoint_complete(dest_key)
-            self.ap.ap_ckb('log+vce', f"Current Waypoint complete.")
+            self.ap_ckb('log+vce', f"Current Waypoint complete.")
 
         # Done with waypoints
         if not _abort:
-            self.ap.ap_ckb('log+vce',
+            self.ap_ckb('log+vce',
                            "Waypoint Route Complete, total distance jumped: " + str(self.ap.total_dist_jumped) + "LY")
             self.ap.update_ap_status("Idle")
         else:
-            self.ap.ap_ckb('log+vce', "Waypoint Route was aborted.")
+            self.ap_ckb('log+vce', "Waypoint Route was aborted.")
             self.ap.update_ap_status("Idle")
 
     def reset_stats(self):
@@ -699,7 +700,7 @@ def main():
     from ED_AP import EDAutopilot
 
     ed_ap = EDAutopilot(cb=None)
-    wp = EDWayPoint(ed_ap, True)  # False = Horizons
+    wp = EDWayPoint(ed_ap, None, True)  # False = Horizons
     wp.step = 0  # start at first waypoint
     keys = EDKeys(cb=None)
     keys.activate_window = True
