+++ ED_AP.py
@@ -69,11 +69,6 @@ class FSDAssistReturn(Enum):
     Complete = 2  # Reached final system and there is no local destination
 
 
-def scale(inp: float, in_min: float, in_max: float, out_min: float, out_max: float) -> float:
-    """ Does scaling of the input based on input and output min/max."""
-    return (inp - in_min)/(in_max - in_min) * (out_max - out_min) + out_min
-
-
 class EDAutopilot:
 
     def __init__(self, cb, doThread=True):
@@ -1479,11 +1474,13 @@ class EDAutopilot:
     def is_sun_dead_ahead(self, scr_reg):
         return scr_reg.sun_percent(scr_reg.screen) > 5
 
-    # use to orient the ship to not be pointing right at the Sun
-    # Checks brightness in the region in front of us, if brightness exceeds a threshold
-    # then will pitch up until below threshold.
-    #
     def sun_avoid(self, scr_reg):
+        """ Use to orient the ship to not be pointing right at the Sun
+        Checks brightness in the region in front of us, if brightness exceeds a threshold
+        then will pitch up until below threshold.
+        @param scr_reg:
+        @return:
+        """
         logger.debug('align= avoid sun')
 
         sleep(0.5)
@@ -1986,129 +1983,6 @@ class EDAutopilot:
         raise Exception("FSD Jump failure")
 
         # a set of convience routes to pitch, rotate by specified degress
-
-    def roll_clockwise_anticlockwise(self, deg):
-        abs_deg = abs(deg)
-        htime = abs_deg/self.rollrate
-
-        if self.speed_demand is None:
-            self.set_speed_50()
-
-        # Calculate rate for less than 45 degrees, else use default
-        if self.current_ship_cfg:
-            if abs_deg < 45:
-                # Roll rate from ship config
-                if self.speed_demand not in self.current_ship_cfg:
-                    self.current_ship_cfg[self.speed_demand] = dict()
-                speed_demand = self.current_ship_cfg[self.speed_demand]
-                if 'RollRate' not in speed_demand:
-                    speed_demand['RollRate'] = dict()
-
-                last_deg = 0.0
-                last_val = 0.0
-                for key, value in speed_demand['RollRate'].items():
-                    key_deg = float(int(key)) / 10
-                    if abs_deg <= key_deg:
-                        print(f"Roll demand: {deg}. Closest lookup: {key_deg}, {value}")
-
-                        # Ratio based on the last value and this value
-                        ratio_val = scale(abs_deg, last_deg, key_deg, last_val, value)
-                        print(f"Roll demand: {deg}. Ratio value: {round(ratio_val, 2)}")
-
-                        htime = abs_deg / ratio_val
-                        break
-                    else:
-                        last_deg = key_deg
-                        last_val = value
-
-        # Check if we are rolling right or left
-        if deg > 0.0:
-            self.keys.send('RollRightButton', hold=htime)
-        else:
-            self.keys.send('RollLeftButton', hold=htime)
-
-    def pitch_up_down(self, deg):
-        abs_deg = abs(deg)
-        htime = abs_deg/self.pitchrate
-
-        if self.speed_demand is None:
-            self.set_speed_50()
-
-        # Calculate rate for less than 30 degrees, else use default
-        if self.current_ship_cfg:
-            if abs_deg < 30:
-                # Pitch rate from ship config
-                if self.speed_demand not in self.current_ship_cfg:
-                    self.current_ship_cfg[self.speed_demand] = dict()
-                speed_demand = self.current_ship_cfg[self.speed_demand]
-                if 'PitchRate' not in speed_demand:
-                    speed_demand['PitchRate'] = dict()
-
-                last_deg = 0.0
-                last_val = 0.0
-                for key, value in speed_demand['PitchRate'].items():
-                    key_deg = float(int(key)) / 10
-                    if abs_deg <= key_deg:
-                        print(f"Pitch demand: {deg}. Closest lookup: {key_deg}, {value}")
-
-                        # Ratio based on the last value and this value
-                        ratio_val = scale(abs_deg, last_deg, key_deg, last_val, value)
-                        print(f"Pitch demand: {deg}. Ratio value: {round(ratio_val, 2)}")
-
-                        htime = abs_deg / ratio_val
-                        break
-                    else:
-                        last_deg = key_deg
-                        last_val = value
-
-        # Check if we are pitching up or down
-        if deg > 0.0:
-            self.keys.send('PitchUpButton', hold=htime)
-        else:
-            self.keys.send('PitchDownButton', hold=htime)
-
-    def yaw_right_left(self, deg):
-        """ Yaw in deg. (> 0.0 for yaw right, < 0.0 for yaw left)
-        @return: The key hold duration.
-        """
-        abs_deg = abs(deg)
-        htime = abs_deg/self.yawrate
-
-        if self.speed_demand is None:
-            self.set_speed_50()
-
-        # Calculate rate for less than 30 degrees, else use default
-        if self.current_ship_cfg:
-            if abs_deg < 30:
-                # Yaw rate from ship config
-                if self.speed_demand not in self.current_ship_cfg:
-                    self.current_ship_cfg[self.speed_demand] = dict()
-                speed_demand = self.current_ship_cfg[self.speed_demand]
-                if 'YawRate' not in speed_demand:
-                    speed_demand['YawRate'] = dict()
-
-                last_deg = 0.0
-                last_val = 0.0
-                for key, value in speed_demand['YawRate'].items():
-                    key_deg = float(int(key)) / 10
-                    if abs_deg <= key_deg:
-                        print(f"Yaw demand: {deg}. Closest lookup: {key_deg}, {value}")
-
-                        # Ratio based on the last value and this value
-                        ratio_val = scale(abs_deg, last_deg, key_deg, last_val, value)
-                        print(f"Yaw demand: {deg}. Ratio value: {round(ratio_val, 2)}")
-
-                        htime = abs_deg / ratio_val
-                        break
-                    else:
-                        last_deg = key_deg
-                        last_val = value
-
-        # Check if we are yawing right or left
-        if deg > 0.0:
-            self.keys.send('YawRightButton', hold=htime)
-        else:
-            self.keys.send('YawLeftButton', hold=htime)
 
     def refuel(self, scr_reg):
         """ Check if refueling needed, ensure correct start type. """
@@ -3035,279 +2909,6 @@ class EDAutopilot:
             self.update_overlay()
             cv2.waitKey(10)
             sleep(1)
-
-    def ship_tst_pitch(self, angle: float):
-        """ Performs a ship pitch test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the pitch value.
-        If the ship rotates too much, increase the pitch value.
-        """
-        # if not self.status.get_flag(FlagsSupercruise):
-        #     self.ap_ckb('log', "Enter Supercruise and try again.")
-        #     return
-        #
-        # if self.jn.ship_state()['target'] is None:
-        #     self.ap_ckb('log', "Select a target system and try again.")
-        #     return
-
-        set_focus_elite_window()
-        sleep(0.25)
-        # self.set_speed_50()
-        self.pitch_up_down(angle)
-
-    def ship_tst_pitch_new(self, angle: float):
-        """ Performs a ship pitch test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the pitch value.
-        If the ship rotates too much, increase the pitch value.
-        """
-        self.ap_ckb('log', "Starting Pitch Calibration.")
-        if not self.speed_demand == 'SCSpeed50':
-            self.set_speed_50()
-            # sleep(10)
-
-        if not self.current_ship_cfg:
-            return
-
-        if self.speed_demand not in self.current_ship_cfg:
-            self.current_ship_cfg[self.speed_demand] = dict()
-
-        # Clear existing data
-        self.current_ship_cfg[self.speed_demand]['PitchRate'] = dict()
-
-        test_time = 0.05
-        delta_int = 0.0
-        for targ_ang in [5, 10, 20, 40, 80, 160]:
-            while 1:
-                set_focus_elite_window()
-                off = self.get_target_offset(self.scrReg)
-                if not off:
-                    print(f"Target lost")
-                    break
-
-                # Clear the overlays before moving
-                if self.debug_overlay:
-                    self.overlay.overlay_remove_rect('target')
-                    self.overlay.overlay_remove_floating_text('target')
-                    self.overlay.overlay_remove_floating_text('target_occ')
-                    self.overlay.overlay_remove_floating_text('target_rpy')
-                    self.overlay.overlay_paint()
-
-                if off['pit'] > 0:
-                    self.keys.send('PitchUpButton', hold=test_time)
-                else:
-                    self.keys.send('PitchDownButton', hold=test_time)
-
-                sleep(1)
-
-                off2 = self.get_target_offset(self.scrReg)
-                if not off2:
-                    print(f"Target lost")
-                    break
-
-                delta = abs(off2['pit'] - off['pit'])
-                delta_int_lst = delta_int
-                delta_int = int(round(delta * 10, 0))
-
-                test_time = test_time * 1.04
-                rate = round(delta / test_time, 2)
-                rate = min(rate, self.pitchrate)  # Limit rate to no higher than the default
-                if delta_int >= targ_ang and delta_int > delta_int_lst:
-                    self.current_ship_cfg[self.speed_demand]['PitchRate'][delta_int] = rate
-
-                    print(f"Pitch Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    self.ap_ckb('log', f"Pitch Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    break
-                else:
-                    print(f"Ignored Pitch Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-
-        # If we have logged values, add the ship default rate at 30 deg
-        if len(self.current_ship_cfg[self.speed_demand]['PitchRate']) > 0:
-            self.current_ship_cfg[self.speed_demand]['PitchRate'][300] = self.pitchrate
-            self.ap_ckb('log', f"Default: Pitch Angle: 30: Rate: {self.pitchrate}")
-
-        self.ap_ckb('log', "Completed Pitch Calibration.")
-        self.ap_ckb('log', "Remember to Save if you wish to keep these values!")
-
-    def ship_tst_roll(self, angle: float):
-        """ Performs a ship roll test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the roll value.
-        If the ship rotates too much, increase the roll value.
-        """
-        # if not self.status.get_flag(FlagsSupercruise):
-        #     self.ap_ckb('log', "Enter Supercruise and try again.")
-        #     return
-        #
-        # if self.jn.ship_state()['target'] is None:
-        #     self.ap_ckb('log', "Select a target system and try again.")
-        #     return
-
-        set_focus_elite_window()
-        sleep(0.25)
-        # self.set_speed_50()
-        self.roll_clockwise_anticlockwise(angle)
-
-    def ship_tst_roll_new(self, angle: float):
-        """ Performs a ship roll test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the roll value.
-        If the ship rotates too much, increase the roll value.
-        """
-        self.ap_ckb('log', "Starting Roll Calibration.")
-        if not self.speed_demand == 'SCSpeed50':
-            self.set_speed_50()
-            #sleep(10)
-
-        if not self.current_ship_cfg:
-            return
-
-        if self.speed_demand not in self.current_ship_cfg:
-            self.current_ship_cfg[self.speed_demand] = dict()
-
-        # Clear existing data
-        self.current_ship_cfg[self.speed_demand]['RollRate'] = dict()
-
-        test_time = 0.05
-        delta_int = 0.0
-        for targ_ang in [40, 80, 160, 320]:
-            while 1:
-                set_focus_elite_window()
-                off = self.get_nav_offset(self.scrReg)
-                if not off:
-                    break
-
-                # Clear the overlays before moving
-                if self.debug_overlay:
-                    self.overlay.overlay_remove_rect('compass')
-                    self.overlay.overlay_remove_floating_text('compass')
-                    self.overlay.overlay_remove_floating_text('nav')
-                    self.overlay.overlay_remove_floating_text('nav_beh')
-                    self.overlay.overlay_remove_floating_text('compass_rpy')
-                    self.overlay.overlay_paint()
-
-                if off['roll'] > 0:
-                    self.keys.send('RollRightButton', hold=test_time)
-                else:
-                    self.keys.send('RollLeftButton', hold=test_time)
-
-                sleep(1)
-
-                off2 = self.get_nav_offset(self.scrReg)
-                if not off2:
-                    break
-
-                delta = abs(off2['roll'] - off['roll'])
-                delta_int_lst = delta_int
-                delta_int = int(round(delta * 10, 0))
-
-                test_time = test_time * 1.03
-                rate = round(delta / test_time, 2)
-                rate = min(rate, self.rollrate)  # Limit rate to no higher than the default
-                if delta_int >= targ_ang and delta_int > delta_int_lst:
-                    self.current_ship_cfg[self.speed_demand]['RollRate'][delta_int] = rate
-
-                    print(f"Roll Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    self.ap_ckb('log', f"Roll Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    break
-                else:
-                    print(f"Ignored Roll Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-
-        # If we have logged values, add the ship default rate at 45 deg
-        if len(self.current_ship_cfg[self.speed_demand]['RollRate']) > 0:
-            self.current_ship_cfg[self.speed_demand]['RollRate'][450] = self.rollrate
-            self.ap_ckb('log', f"Default: Roll Angle: 45: Rate: {self.rollrate}")
-
-        self.ap_ckb('log', "Completed Roll Calibration.")
-        self.ap_ckb('log', "Remember to Save if you wish to keep these values!")
-
-    def ship_tst_yaw(self, angle: float):
-        """ Performs a ship yaw test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the yaw value.
-        If the ship rotates too much, increase the yaw value.
-        """
-        # if not self.status.get_flag(FlagsSupercruise):
-        #     self.ap_ckb('log', "Enter Supercruise and try again.")
-        #     return
-        #
-        # if self.jn.ship_state()['target'] is None:
-        #     self.ap_ckb('log', "Select a target system and try again.")
-        #     return
-
-        set_focus_elite_window()
-        sleep(0.25)
-        # self.set_speed_50()
-        self.yaw_right_left(angle)
-
-    def ship_tst_yaw_new(self, angle: float):
-        """ Performs a ship yaw test by pitching 360 degrees.
-        If the ship does not rotate enough, decrease the yaw value.
-        If the ship rotates too much, increase the yaw value.
-        """
-        self.ap_ckb('log', "Starting Yaw Calibration.")
-        # self.compass_align(self.scrReg)
-        # self.sc_target_align(self.scrReg)
-
-        if not self.speed_demand == 'SCSpeed50':
-            self.set_speed_50()
-            # sleep(10)
-
-        if not self.current_ship_cfg:
-            return
-
-        if self.speed_demand not in self.current_ship_cfg:
-            self.current_ship_cfg[self.speed_demand] = dict()
-
-        # Clear existing data
-        self.current_ship_cfg[self.speed_demand]['YawRate'] = dict()
-
-        test_time = 0.07
-        delta_int = 0.0
-        for targ_ang in [5, 10, 20, 40, 80, 160]:
-            while 1:
-                set_focus_elite_window()
-                off = self.get_target_offset(self.scrReg)
-                if not off:
-                    break
-
-                # Clear the overlays before moving
-                if self.debug_overlay:
-                    self.overlay.overlay_remove_rect('target')
-                    self.overlay.overlay_remove_floating_text('target')
-                    self.overlay.overlay_remove_floating_text('target_occ')
-                    self.overlay.overlay_remove_floating_text('target_rpy')
-                    self.overlay.overlay_paint()
-
-                if off['yaw'] > 0:
-                    self.keys.send('YawRightButton', hold=test_time)
-                else:
-                    self.keys.send('YawLeftButton', hold=test_time)
-
-                sleep(1)
-
-                off2 = self.get_target_offset(self.scrReg)
-                if not off2:
-                    break
-
-                delta = abs(off2['yaw'] - off['yaw'])
-                delta_int_lst = delta_int
-                delta_int = int(round(delta * 10, 0))
-
-                test_time = test_time * 1.05
-                rate = round(delta / test_time, 2)
-                rate = min(rate, self.yawrate)  # Limit rate to no higher than the default
-                if delta_int >= targ_ang and delta_int > delta_int_lst:
-                    self.current_ship_cfg[self.speed_demand]['YawRate'][delta_int] = rate
-
-                    print(f"Yaw Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    self.ap_ckb('log', f"Yaw Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-                    break
-                else:
-                    print(f"Ignored Yaw Angle: {round(delta, 2)}: Time: {round(test_time, 2)} Rate: {rate}")
-
-        # If we have logged values, add the ship default rate at 30 deg
-        if len(self.current_ship_cfg[self.speed_demand]['YawRate']) > 0:
-            self.current_ship_cfg[self.speed_demand]['YawRate'][300] = self.yawrate
-            self.ap_ckb('log', f"Default: Yaw Angle: 30: Rate: {self.yawrate}")
-
-        self.ap_ckb('log', "Completed Yaw Calibration.")
-        self.ap_ckb('log', "Remember to Save if you wish to keep these values!")
 
     def set_speed_0(self, repeat=1):
         if self.status.get_flag(FlagsSupercruise):
